깃 허브 웹훅 로직 수정

+++ external-client/git-hub/src/main/java/com/ryuqq/core/external/git/GitHubFeignClient.java
@@ -18,8 +18,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 public interface GitHubFeignClient {
 
 	@GetMapping("/repos/{fullName}/pulls/{pullNumber}/files")
-	List<GitHubCommitResponse> getCommitInfos(@PathVariable("fullName") String fullName,
-									   @PathVariable("pullNumber") int pullNumber);
-
+	List<GitHubCommitResponse> getCommitInfos(@PathVariable("fullName") String fullName, @PathVariable("pullNumber") long pullNumber);
 
 }
